Crate picotoml

A no_std toml deserializer for embedded systems Full toml support for deserializing into structs that implement serde::Deserialize.

Requires no memory allocations and is likely much faster than toml-rs. Supports deserialization only

Macros

expect_next

expect_next_peeked

expect_next_with_item

range_to_str

Returns a slice containing the text of the given token item using the provided deserializer

Structs

Deserializer

Error

Enums

ErrorKind

Expected

Functions

from_str